+++ sdk/spring/azure-spring-boot/src/test/java/com/azure/spring/autoconfigure/cosmos/CosmosPropertiesTest.java
@@ -3,7 +3,7 @@
 package com.azure.spring.autoconfigure.cosmos;
 
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 import org.springframework.boot.context.properties.ConfigurationPropertiesBindException;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.boot.context.properties.bind.validation.BindValidationException;
@@ -19,6 +19,7 @@ import java.util.stream.Collectors;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class CosmosPropertiesTest {
+
     @Test
     public void canSetAllProperties() {
         PropertySettingUtil.setProperties();
@@ -77,9 +78,9 @@ public class CosmosPropertiesTest {
             Collections.sort(errorStrings);
 
             final List<String> errorStringsExpected = Arrays.asList(
-                    "Field error in object 'azure.cosmos' on field 'database': rejected value [null];",
-                    "Field error in object 'azure.cosmos' on field 'key': rejected value [null];",
-                    "Field error in object 'azure.cosmos' on field 'uri': rejected value [null];"
+                "Field error in object 'azure.cosmos' on field 'database': rejected value [null];",
+                "Field error in object 'azure.cosmos' on field 'key': rejected value [null];",
+                "Field error in object 'azure.cosmos' on field 'uri': rejected value [null];"
             );
 
             assertThat(errorStrings.size()).isEqualTo(errorStringsExpected.size());